requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
bidict==0.23.1
billiard==4.2.1
celery==5.5.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
Django==5.1.4
django-autocomplete-light==3.11.0
django-cors-headers==4.9.0
django-countries==8.2.0
django-filter==24.3
django-ipware==7.0.1
django-redis==6.0.0
django-tables2==2.7.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
eventlet==0.40.2
gevent==25.5.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
hashids==1.3.1
hiredis==3.3.0
idna==3.10
intasend-python==1.1.2
kombu==5.5.4
ndg-httpsclient==0.5.1
packaging==24.2
phonenumbers==8.13.52
prompt_toolkit==3.0.51
psycogreen==1.0.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-ipware==3.0.0
pytz==2025.2
redis==6.4.0
requests==2.32.3
schema==0.7.7
sentry-sdk==2.48.0
setuptools==82.0.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
